- RealmList Parcelable 처리한거 주석 추가하여 util로 이동 - 금액 포맷 추가 - history 레이아웃 수정
