+++ app/src/main/res/drawable/bg_tag_rounded.xml
@@ -7,7 +7,7 @@
         <shape android:shape="rectangle">
             <solid android:color="@color/backgroundColor" />
             <stroke android:color="@color/light_grey" android:width="1dp" />
-            <corners android:radius="16dp" />
+            <corners android:radius="20dp" />
         </shape>
     </item>
